Move the authentication activies to their own package
