+++ Simplenote/src/main/AndroidManifest.xml
@@ -121,14 +121,14 @@
         </activity>
 
         <activity
-            android:name="com.automattic.simplenote.SimplenoteAuthenticationActivity">
+            android:name="com.automattic.simplenote.authentication.SimplenoteAuthenticationActivity">
         </activity>
 
         <activity
-            android:name="com.automattic.simplenote.SimplenoteCredentialsActivity">
+            android:name="com.automattic.simplenote.authentication.SimplenoteCredentialsActivity">
         </activity>
 
-        <activity android:name=".SimplenoteSignupActivity" />
+        <activity android:name=".authentication.SimplenoteSignupActivity" />
 
         <activity
             android:name="com.automattic.simplenote.StyleActivity"
