+++ Simplenote/src/main/java/com/automattic/simplenote/authentication/SimplenoteSignupActivity.java
@@ -1,9 +1,11 @@
-package com.automattic.simplenote;
+package com.automattic.simplenote.authentication;
 
 import android.os.Bundle;
 
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.automattic.simplenote.R;
 
 public class SimplenoteSignupActivity extends AppCompatActivity {
     @Override
